application page edits
