--- client/src/pages/OrgWorkSpace.js
@@ -1,563 +1,0 @@
-import { WhiteSpace } from "antd-mobile";
-import { Tabs } from "antd";
-import { ProfileTabs, ProfileTabPane } from "components/OrganisationProfile/ProfileTabs"
-import axios from "axios";
-import React, {
-    useState,
-    useEffect,
-    useContext,
-    useCallback,
-    useReducer,
-    useRef,
-} from "react";
-import { useSelector, useDispatch } from "react-redux";
-import { Link } from "react-router-dom";
-import { useTranslation } from "react-i18next";
-
-// ICONS
-import createPost from "assets/icons/create-post.svg";
-import edit from "assets/icons/edit.svg";
-import locationIcon from "assets/icons/location.svg";
-import envelopeBlue from "assets/icons/social-envelope-blue.svg";
-import playStoreIcon from "assets/icons/play-store-icon.svg";
-import appStoreIcon from "assets/icons/app-store-icon.svg";
-
-import instagramIcon from "assets/icons/social-instagram.svg";
-import linkedinBlue from "assets/icons/social-linkedin.svg";
-import facebookIcon from "assets/icons/social-fb.svg";
-import twitterBlue from "assets/icons/social-tw.svg";
-import githubIcon from "assets/icons/social-github.svg";
-import websiteIcon from "assets/icons/website-icon.svg";
-
-import Activity from "components/Profile/Activity";
-import CreatePost from "components/CreatePost/CreatePost";
-import ErrorAlert from "../components/Alert/ErrorAlert";
-import { FeedWrapper } from "components/Feed/FeedWrappers";
-import ProfilePic from "components/Picture/ProfilePic";
-import UploadPic from "components/Picture/UploadPic";
-import MessageModal from "../components/Feed/MessagesModal/MessageModal.js";
-import Verification from "components/Verification/";
-import VerificationTick from "components/Verification/Tick";
-
-import Loader from "components/Feed/StyledLoader";
-import {
-    ProfileLayout,
-    UserInfoContainer,
-    EditIcon,
-    UserInfoDesktop,
-    NameDiv,
-    PlaceholderIcon,
-    DescriptionDesktop,
-    IconsContainer,
-    SocialIcon,
-    SectionHeader,
-    CreatePostDiv,
-    CreatePostIcon,
-    DrawerHeader,
-    CustomDrawer,
-    PhotoUploadButton,
-    AvatarPhotoContainer,
-    NamePara,
-    ProfileBackgroup,
-} from "../components/Profile/ProfileComponents";
-import {
-    getInitialsFromFullName,
-    isAuthorOrg,
-    isAuthorUser,
-} from "utils/userInfo";
-import {
-    FACEBOOK_URL,
-    INSTAGRAM_URL,
-    LINKEDIN_URL,
-    TWITTER_URL,
-    GITHUB_URL,
-    APPSTORE_URL,
-    PLAYSTORE_URL,
-} from "constants/urls";
-import {
-    fetchOrganisation,
-    fetchOrganisationError,
-    fetchOrganisationSuccess,
-} from "hooks/actions/organisationActions";
-import {
-    fetchUser,
-    fetchUserError,
-    fetchUserSuccess,
-} from "hooks/actions/userActions";
-import {
-    OrganisationContext,
-    withOrganisationContext,
-} from "context/OrganisationContext";
-import { SET_EDIT_POST_MODAL_VISIBILITY } from "hooks/actions/postActions";
-import {
-    SET_DELETE_MODAL_VISIBILITY,
-    DELETE_MODAL_POST,
-    DELETE_MODAL_HIDE,
-} from "hooks/actions/feedActions";
-import {
-    deletePostModalreducer,
-    deletePostState,
-} from "hooks/reducers/feedReducers";
-import { UserContext, withUserContext } from "context/UserContext";
-import GTM from "constants/gtm-tags";
-import { selectPosts, postsActions } from "reducers/posts";
-import { selectOrganisationId } from "reducers/session";
-import CreatePostButton from "components/Feed/CreatePostButton";
-import { ReactComponent as PlusIcon } from "assets/icons/pretty-plus.svg";
-import JoinOrgButton, { JoinOrgContainer } from "components/OrganisationProfile/JoinOrgButton";
-import { LOGIN } from "templates/RouteWithSubRoutes";
-import TestMembers from "components/OrganisationProfile/ProfileList"
-
-const URLS = {
-    playStore: [playStoreIcon, PLAYSTORE_URL],
-    appStore: [appStoreIcon, APPSTORE_URL],
-    facebook: [facebookIcon, FACEBOOK_URL],
-    instagram: [instagramIcon, INSTAGRAM_URL],
-    linkedin: [linkedinBlue, LINKEDIN_URL],
-    twitter: [twitterBlue, TWITTER_URL],
-    github: [githubIcon, GITHUB_URL],
-    website: [websiteIcon],
-    email: [envelopeBlue],
-};
-
-const getHref = (url) => (url.startsWith("http") ? url : `//${url}`);
-const PAGINATION_LIMIT = 10;
-const ARBITRARY_LARGE_NUM = 10000;
-const OrgWorkSpace = ({ isAuthenticated }) => {
-    let url = window.location.pathname.split("/");
-    const organisationId = url[url.length - 1];
-    const { orgProfileState, orgProfileDispatch } = useContext(
-        OrganisationContext,
-    );
-    const { error, loading, organisation } = orgProfileState;
-    const [deleteModal, deleteModalDispatch] = useReducer(
-        deletePostModalreducer,
-        deletePostState,
-    );
-    const posts = useSelector(selectPosts);
-
-    const {
-        userProfileState: { user },
-        userProfileDispatch,
-    } = useContext(UserContext);
-
-    const { t } = useTranslation();
-    const dispatch = useDispatch();
-    const [modal, setModal] = useState(false);
-    const [drawer, setDrawer] = useState(false);
-    const [itemCount, setItemCount] = useState(0);
-    const [toggleRefetch, setToggleRefetch] = useState(false);
-    const [totalPostCount, setTotalPostCount] = useState(ARBITRARY_LARGE_NUM);
-    const {
-        email,
-        name,
-        location = {},
-        about = "",
-        isOwner,
-        urls = {},
-        verified,
-    } = organisation || {};
-
-    const urlsAndEmail = { ...urls, email: isOwner ? null : email };
-    if (isOwner) sessionStorage.removeItem("msgModal");
-    const {
-        isLoading,
-        loadMore,
-        page,
-        posts: postsList,
-        error: postsError,
-    } = posts;
-    const { deleteModalVisibility } = deleteModal;
-
-    const prevTotalPostCount = usePrevious(totalPostCount);
-    const prevOrgId = usePrevious(organisationId);
-    const organisationPosts = Object.entries(postsList);
-    const actorOrganisationId = useSelector(selectOrganisationId);
-    const isSelf = organisation && actorOrganisationId == organisation._id;
-
-    function usePrevious(value) {
-        const ref = useRef();
-        useEffect(() => {
-            ref.current = value;
-        });
-        return ref.current;
-    }
-    const getActorQuery = () => {
-        return actorOrganisationId ? `&actorId=${actorOrganisationId}` : "";
-    };
-
-    useEffect(() => {
-        dispatch(postsActions.resetPageAction({}));
-        (async function fetchOrgProfile() {
-            orgProfileDispatch(fetchOrganisation());
-            userProfileDispatch(fetchUser());
-            try {
-                const res = await axios.get(`/api/organisations/${organisationId}`);
-                orgProfileDispatch(fetchOrganisationSuccess(res.data));
-            } catch (err) {
-                const message = err.response?.data?.message || err.message;
-                const translatedErrorMessage = t([
-                    `error.${message}`,
-                    `error.http.${message}`,
-                ]);
-                orgProfileDispatch(
-                    fetchOrganisationError(
-                        `${t("error.failedLoadingProfile")} ${translatedErrorMessage}`,
-                    ),
-                );
-            }
-        })();
-        (async function fetchUserProfile() {
-            userProfileDispatch(fetchUser());
-            try {
-                const res = await axios.get("/api/users/current");
-                userProfileDispatch(fetchUserSuccess(res.data));
-            } catch (err) {
-                const message = err.response?.data?.message || err.message;
-                const translatedErrorMessage = t([
-                    `error.${message}`,
-                    `error.http.${message}`,
-                ]);
-                userProfileDispatch(
-                    fetchUserError(
-                        `${t("error.failedLoadingProfile")} ${translatedErrorMessage}`,
-                    ),
-                );
-            }
-        })();
-    }, [orgProfileDispatch, organisationId, userProfileDispatch]); // eslint-disable-line react-hooks/exhaustive-deps
-
-    useEffect(() => {
-        const fetchOrganisationPosts = async () => {
-            const limit = PAGINATION_LIMIT;
-            const skip = page * limit;
-            dispatch(postsActions.fetchPostsBegin());
-            try {
-                if (organisationId) {
-                    const endpoint = `/api/posts?ignoreUserLocation=true&includeMeta=true&limit=${limit}&skip=${skip}&authorId=${organisationId}${getActorQuery()}`;
-                    const {
-                        data: { data: posts, meta },
-                    } = await axios.get(endpoint);
-
-                    if (prevOrgId !== organisationId) {
-                        dispatch(
-                            postsActions.fetchPostsSuccess({
-                                posts: [],
-                            }),
-                        );
-                    }
-                    if (posts.length && meta.total) {
-                        if (prevTotalPostCount !== meta.total) {
-                            setTotalPostCount(meta.total);
-                        }
-                        if (posts.length < limit) {
-                            dispatch(postsActions.finishLoadingAction());
-                        } else if (meta.total === limit) {
-                            dispatch(postsActions.finishLoadingAction());
-                        }
-                        const loadedPosts = posts.reduce((obj, item) => {
-                            obj[item._id] = item;
-                            return obj;
-                        }, {});
-
-                        if (prevOrgId === organisationId && postsList) {
-                            dispatch(
-                                postsActions.fetchPostsSuccess({
-                                    posts: { ...postsList, ...loadedPosts },
-                                }),
-                            );
-                        } else {
-                            dispatch(
-                                postsActions.fetchPostsSuccess({
-                                    posts: { ...loadedPosts },
-                                }),
-                            );
-                        }
-                    } else if (prevOrgId === organisationId && posts) {
-                        dispatch(
-                            postsActions.fetchPostsSuccess({
-                                posts: { ...postsList },
-                            }),
-                        );
-                        dispatch(postsActions.finishLoadingAction());
-                    } else {
-                        dispatch(postsActions.finishLoadingAction());
-                    }
-                }
-            } catch (error) {
-                dispatch(postsActions.fetchPostsError(error));
-            }
-        };
-        fetchOrganisationPosts();
-    }, [organisationId, page, toggleRefetch]); // eslint-disable-line react-hooks/exhaustive-deps
-
-    const refetchPosts = (isLoading, loadMore) => {
-        dispatch(postsActions.resetPageAction({ isLoading, loadMore }));
-        if (page === 0) {
-            setToggleRefetch(!toggleRefetch);
-        }
-    };
-
-    const isItemLoaded = useCallback((index) => !!organisationPosts[index], [
-        organisationPosts,
-    ]);
-
-    const loadNextPage = useCallback(
-        ({ stopIndex }) => {
-            if (
-                !isLoading &&
-                loadMore &&
-                stopIndex >= organisationPosts.length &&
-                organisationPosts.length
-            ) {
-                return new Promise((resolve) => {
-                    dispatch(postsActions.setNextPageAction());
-                    resolve();
-                });
-            } else {
-                return Promise.resolve();
-            }
-        },
-        [dispatch, isLoading, loadMore, organisationPosts.length],
-    );
-
-    useEffect(() => {
-        setItemCount(
-            loadMore ? organisationPosts.length + 1 : organisationPosts.length,
-        );
-    }, [loadMore, organisationPosts.length]);
-
-    const postDelete = async (post) => {
-        let deleteResponse;
-        const endPoint = `/api/posts/${post._id}`;
-        if (
-            user &&
-            (isAuthorUser(user, post) || isAuthorOrg(user.organisations, post.author))
-        ) {
-            try {
-                deleteResponse = await axios.delete(endPoint);
-                if (deleteResponse && deleteResponse.data.success === true) {
-                    const allPosts = {
-                        ...postsList,
-                    };
-                    delete allPosts[post._id];
-                    setTotalPostCount(totalPostCount - 1);
-                    if (totalPostCount <= PAGINATION_LIMIT) {
-                        const isLoading = true;
-                        const loadMore = false;
-                        refetchPosts(isLoading, loadMore);
-                    } else {
-                        refetchPosts();
-                    }
-                }
-            } catch (error) {
-                console.log({
-                    error,
-                });
-            }
-        }
-    };
-
-    const handlePostDelete = () => {
-        deleteModalDispatch({
-            type: SET_DELETE_MODAL_VISIBILITY,
-            visibility: DELETE_MODAL_POST,
-        });
-    };
-
-    const handleCancelPostDelete = () => {
-        deleteModalDispatch({
-            type: SET_DELETE_MODAL_VISIBILITY,
-            visibility: DELETE_MODAL_HIDE,
-        });
-    };
-
-    const handleEditPost = () => {
-        if (deleteModal.editPostModalVisibility) {
-            deleteModalDispatch({
-                type: SET_EDIT_POST_MODAL_VISIBILITY,
-                visibility: false,
-            });
-        } else {
-            deleteModalDispatch({
-                type: SET_EDIT_POST_MODAL_VISIBILITY,
-                visibility: true,
-            });
-        }
-    };
-    const renderURL = () => {
-        if (organisation) {
-            if (urlsAndEmail.length !== 0) {
-                return Object.entries(urlsAndEmail).map(([name, url]) => {
-                    let href;
-                    if (name === "website") {
-                        href = getHref(url);
-                    } else if (name === "email") {
-                        href = `mailto:${url}`;
-                    } else {
-                        href = `${URLS[name][1]}${url}`;
-                    }
-
-                    return (
-                        url && (
-                            <a
-                                href={href}
-                                key={name}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                            >
-                                <SocialIcon
-                                    src={URLS[name][0]}
-                                    alt={name}
-                                    id={
-                                        name === "email"
-                                            ? GTM.organisation.orgPrefix + GTM.organisation.email
-                                            : ""
-                                    }
-                                />
-                            </a>
-                        )
-                    );
-                });
-            } else {
-                return;
-            }
-        }
-    };
-
-    const emptyFeed = () => Object.keys(postsList).length < 1 && !isLoading;
-    const onToggleDrawer = () => setDrawer(!drawer);
-    const onToggleCreatePostDrawer = () => setModal(!modal);
-    const { TabPane } = Tabs
-
-    if (error) {
-        return <ErrorAlert message={error} type="error" />;
-    }
-    if (loading) return <Loader />;
-
-    if (!organisation) {
-        return <Loader />;
-    } else {
-        const { address } = location;
-        return (
-            <>
-                <ProfileBackgroup />
-                <ProfileLayout>
-                    <UserInfoContainer>
-                        <AvatarPhotoContainer>
-                            <ProfilePic
-                                user={organisation}
-                                initials={getInitialsFromFullName(name)}
-                            />
-                            <PhotoUploadButton>
-                                {isSelf && (
-                                    <UploadPic
-                                        gtmPrefix={GTM.organisation.orgPrefix}
-                                        user={organisation}
-                                    />
-                                )}
-                            </PhotoUploadButton>
-                        </AvatarPhotoContainer>
-                        <UserInfoDesktop>
-                            <NameDiv>
-                                <div className="name-container">
-                                    <NamePara>
-                                        {name} {verified && <VerificationTick />}
-                                    </NamePara>
-                                    {address && (
-                                        <div title={address} className="address-container">
-                                            <img src={locationIcon} alt={address} />
-                                            {address}
-                                        </div>
-                                    )}
-                                </div>
-                                {isSelf && (
-                                    <EditIcon
-                                        src={edit}
-                                        id={GTM.organisation.orgPrefix + GTM.profile.modify}
-                                        onClick={onToggleDrawer}
-                                    />
-                                )}
-                                {!isOwner &&
-                                    !/Sourced by FightPandemics\ \(.*?\)/.test(name) && (
-                                        <MessageModal
-                                            isAuthenticated={isAuthenticated}
-                                            isFromProfile={true}
-                                            isFromUserCard={"ORG"}
-                                            postAuthorName={name}
-                                            authorId={organisationId}
-                                        />
-                                    )}
-                            </NameDiv>
-                            {about && <DescriptionDesktop> {about} </DescriptionDesktop>}
-
-                            {/* <IconsContainer>
-                <div className="social-icons">{renderURL()}</div>
-              </IconsContainer> */}
-
-                        </UserInfoDesktop>
-                    </UserInfoContainer>
-
-                    {isSelf && !verified && <Verification />}
-                    <WhiteSpace />
-                    {// Only show JoinOrgButton if user is not Member, Wiki Editor, or Admin
-                    }
-
-                    {!isOwner ? <JoinOrgContainer>
-                        <Link
-                            onClick={
-                                () => sessionStorage.setItem("postredirect", window.location.pathname)
-                            }
-                            to={isAuthenticated ? `/organisation/${organisationId}/positions` :
-                                {
-                                    pathname: LOGIN,
-                                    state: { from: window.location.pathname },
-                                }}>
-                            <JoinOrgButton
-                                id={GTM.organisation.joinOrg}>
-                                {t("profile.individual.joinOrg")}
-                            </JoinOrgButton>
-                        </Link>
-                    </JoinOrgContainer> : null}
-                    {// TABS
-                    }
-                    <ProfileTabs defaultActiveKey="members">
-                        <ProfileTabPane tab={t("profile.views.members")} key="members"><TestMembers /></ProfileTabPane>
-                    </ProfileTabs>
-
-                    {/* {isSelf && (
-                        <CustomDrawer
-                            placement="bottom"
-                            closable={false}
-                            onClose={onToggleDrawer}
-                            visible={drawer}
-                            height="auto"
-                            key="bottom"
-                        >
-                            <DrawerHeader>
-                                <Link to={`/edit-organisation-account/${organisationId}`}>
-                                    {t("profile.org.editOrgAccount")}
-                                </Link>
-                            </DrawerHeader>
-                            <DrawerHeader>
-                                <Link to={`/edit-organisation-profile/${organisationId}`}>
-                                    {t("profile.org.editOrgProfile") + " "}
-                                </Link>
-                            </DrawerHeader>
-                            <DrawerHeader>
-                                <Link to={`/edit-organisation-notifications/${organisationId}`}>
-                                    {t("profile.org.editOrgNotification")}{" "}
-                                </Link>
-                            </DrawerHeader>
-                        </CustomDrawer>
-                    )} */}
-
-                </ProfileLayout>
-
-
-            </>
-        );
-    }
-};
-
-export default withUserContext(withOrganisationContext(OrgWorkSpace));
